requirements.txt
fastapi
uvicorn
faiss-cpu
tqdm
together
python-dotenv
pypdf
sentence-transformers
